nn_calibration/requirements.txt
numpy>=2.4.3
pytest>=9.0.2
pyYAML>=6.0.3
scipy>=1.17.1
scikit-learn>=1.8.0
